Code is finally done, cleaned up the code too
I fixed everything with the rolling of items code, it was not fun. Had a lot of problems but it was because I was calling things from places I couldn't access certain codes.

I "cleaned" the code up a little bit and added somne comments to explain my thought process.

Had to use these websites to fix my code:
https://www.w3schools.com/java/
https://stackoverflow.com/questions/tagged/java
 Changes to be committed:
	modified:   Main.java
	modified:   playPackage/Play.java